Basic cursor-following of paddle along sphere.
Need to figure out why perspective is strange. On the boundary between
the sphere and the plane, the paddle doesn't actually approach the plane
(z-value of 0); rather, the farthest it goes on the sphere is about z=2,
and then it jumps to zero as soon as it hits the plane. It's also
impossible to see the paddle at z=0 when the sphere is visible, because
it's blocked -- need to figure out why.