Implement box-sphere collision check!
Spent way too long on debugging my brain (see line 181 for full
details). Currently lazily still bounces off sphere at point of contact
rather than off the paddle's normal, but will get to that.

Also added a bunch of hacky debugging tools (printing to page, color of
paddle, keyboard controls).

Still need to figure out how to get the best of both worlds in terms of
cameras -- we ideally want the range of movement on the sphere of the
orthographic camera as well as the depth perception of the ball from the
perspective camera. Without one or the other, we either have Skippy
McSkippers near the edge of the sphere or a ninja ball that's impossible
to judge.