Fix several ball-paddle collision issues.
Corrected the reflection of the ball from the paddle to use the paddle's
normal. Still not completely right, as it assumes that it's colliding
from the inner-facing side of the paddle, but in terms of normal
gameplay it should be good for now. May fix later.

Also add a velocity direction check during those collisions to avoid
tunneling issues (aka the jiggly jiggles), as we now only reflect when
the ball is traveling toward the paddle.