Modify collision checking and add pausing.
Spent a long time figuring out how to best do the check on the ball
against all 6 faces of the paddle (kept _distancePointPlane function in
case we want it in the future), but ultimately just changed lines
231-232. Now it just checks against the front-facing plane -- still
wrong, but practically actually more accurate than the old version
because the front face is basically all we need to check against in
normal play. If the ball doesn't hit the front face, it's not staying in
play anyway.

Also add pause functionality. Currently just removes mousemove event
listener (and adds it back upon unpause) and zeroes the ball velocity.
The current paddle-mouse-following functionality is still a bit
troubling -- should we keep it as it is (event listener), or should we
do the reverse and update the position every loop? Normally the event
listener would be much less wasteful, but in the game the mouse should
be constantly moving anyway, and this would fix the problem where the
paddle doesn't appear in the right spot until the mouse is moved (i.e.
in the beginning of the game or after an unpause). Will think about it.