Tweak dotted lines.
Whoops! Realized that since the second vertices of the ball- and
player-following lines were literally the position vectors, I didn't
have to manually update them at every turn.

Also added a circle to show the radius of the projected point of the
ball. Might do this with the player too. However, with all these
additions, each new thing seems to make the performance take a dip,
especially since all the vertices need to be updated. Also the amount of
position judgment helpers is getting ridiculous, and all the additions
still don't seem to improve the experience much. Hm.