Start adding game features!
Putting the camera-sphere range problem on hold for now, and starting
to actually make this a game. Add a score (currently updates every time
the paddle is hit -- whether it should be the paddle or the ball, I
don't know yet) (currently just displayed in the debugText). Also fix
ball position to (0, 0, 0) and randomize ball velocity upon start and
upon reset with 'r'. Also make the ball speed up with every bounce
against the back wall. Currently linearly increases the speed; seems to
work okay.

Also extend the size of the back plane (still not solving the problem of
universally being able to work, but now for general use (without
manipulating the camera) it's big enough).